requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==9.5.0
moviepy==1.0.3


groq==0.9.0
httpx==0.24.1
httpcore==0.17.3
